vegEcom/requirements.txt
asgiref==3.3.4
astroid==2.5.6
cffi==1.14.5
colorama==0.4.4
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2
django-rest-knox==4.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
pylint==2.8.2
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
whitenoise==5.2.0
wrapt==1.12.1
